backend/requirements.txt
# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# 数据库
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL 驱动

# 数据验证
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP客户端
httpx>=0.25.0

# AI/LLM
openai>=1.7.0

# 定时任务
apscheduler>=3.10.0

# 重试机制
tenacity>=8.2.0

# 日志
structlog>=23.2.0

# HTML解析
beautifulsoup4>=4.12.0
lxml>=4.9.0
html2text>=2024.2.26

# 配置文件
pyyaml>=6.0.0

# RSS解析
feedparser>=6.0.0

# 时间处理
python-dateutil>=2.8.0

# 浏览器自动化
playwright>=1.40.0

# 缓存
redis>=5.0.0  # Redis 缓存（原生支持 async，不需要 aioredis）

# 阿里云SDK
aliyun-python-sdk-core>=2.14.0

# 认证和安全
python-jose[cryptography]>=3.3.0  # JWT处理
passlib[bcrypt]>=1.7.4  # 密码哈希（可选）

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # 用于TestClient

# 类型检查
mypy>=1.7.0
types-redis>=4.6.0

# 代码质量
black>=23.12.0
ruff>=0.1.0
isort>=5.13.0
